[level 2] Title: 멀리 뛰기, Time: 1.79 ms, Memory: 31.2 MB -BaekjoonHub

+++ 프로그래머스/lv2/12914. 멀리 뛰기/멀리 뛰기.cs
@@ -1,14 +1,14 @@
 using System.Collections.Generic;
 public class Solution {
-    Dictionary<int, long> dic = new Dictionary<int, long>();
-    public long solution(int n) {    
+    Dictionary<int, int> dic = new Dictionary<int, int>();
+    public int solution(int n) {    
         dic.Add(1,1); 
         dic.Add(2,2); 
             
         return fibo(n);    
     }
     
-    long fibo(int n){
+    int fibo(int n){
         if (dic.ContainsKey(n) == false)
             dic.Add(n, fibo(n-2) + fibo(n-1));
         return dic[n] % 1234567;
